+++ phono3py/conductivity/__init__.py
@@ -1,0 +1,1 @@
+"""Thermal conductivity related modules."""
